+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- add option to filter by primary source to merged and preview endpoint
+
 ### Changes
 
 ### Deprecated